Extract the `ExternalNetworkMode` enum out of `cuttlefish_config.h`
https://github.com/google/android-cuttlefish/blob/d748b8ae5b18ccb5d0456c8ccc05da3682e0553a/base/cvd/cuttlefish/host/libs/config/cuttlefish_config.h#L49

Bug: b/435771697
