+++ base/cvd/cuttlefish/host/libs/config/BUILD.bazel
@@ -105,6 +105,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:config_fragment",
         "//cuttlefish/host/libs/config:config_utils",
+        "//cuttlefish/host/libs/config:external_network_mode",
         "//cuttlefish/host/libs/config:secure_hals",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
@@ -162,6 +163,17 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:subprocess",
         "//cuttlefish/host/libs/config:known_paths",
         "//libbase",
+    ],
+)
+
+cf_cc_library(
+    name = "external_network_mode",
+    srcs = ["external_network_mode.cc"],
+    hdrs = ["external_network_mode.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+        "@fmt",
     ],
 )
