+++ base/cvd/cuttlefish/host/libs/vm_manager/BUILD.bazel
@@ -45,6 +45,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/command_util",
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:external_network_mode",
         "//cuttlefish/host/libs/config:known_paths",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/feature",
